feat(weather): Add new Weather condition
- add SnowWeatherEffect with SnowFlake,
- add RAinWeatherEffect with Drop,
- add CloudyWeatherEffect with Cloud.

+++ src/main/java/tutorials/weather/WeatherEffect.java
@@ -6,7 +6,7 @@ public interface WeatherEffect {
 
     void draw(Graphics2D g);
 
-    void update(double elapsed);
+    void update(long elapsed);
 
     boolean isActive();
 }
